// src/app/models/image.model.ts
export class ImageModel {
	id: number;
	name: string;
	road: string;
	dateTaken: string;
	dateUploaded: string;
	ido: string;
	keido: string;
	url: any;

	constructor() {
		this.id = undefined;
		this.name = '';
		this.road = '';
		this.dateTaken = '';
		this.dateTaken = '';
		this.ido = '';
		this.keido = '';
		this.url = '';
	}
}